Separate ARF ensemble code from ARF Hoeffding Tree code.
